Fix file.rmcontents() to work when directory is missing.
It would crash otherwise if the directory doesn't exist.

R=phajdan.jr@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1148993004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295346 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/recipe_modules/file/api.py
@@ -169,11 +169,13 @@ class FileApi(recipe_api.RecipeApi):
       import os, sys
       from common import chromium_utils
 
-      for p in [os.path.join(sys.argv[1], x) for x in os.listdir(sys.argv[1])]:
-        if os.path.isdir(p):
-          chromium_utils.RemoveDirectory(p)
-        else:
-          os.unlink(p)
+      path = sys.argv[1]
+      if os.path.exists(path):
+        for p in (os.path.join(path, x) for x in os.listdir(path)):
+          if os.path.isdir(p):
+            chromium_utils.RemoveDirectory(p)
+          else:
+            os.unlink(p)
       """,
       args=[path],
       **kwargs
